% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{plot a heatmap using ComplexHeatmap}
\usage{
draw_heatmap(
  nes_mat,
  q_mat,
  col_df = NULL,
  group,
  cellwidth = 15,
  cellheight = 12
)
}
\description{
plot a heatmap using ComplexHeatmap
}
