% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_case-utils.R
\name{gsea_result}
\alias{gsea_result}
\alias{gsea_result.gsea_case}
\title{Extract the GSEA results table}
\usage{
gsea_result(x, ...)

\method{gsea_result}{gsea_case}(x, ...)
}
\arguments{
\item{x}{An object with GSEA results.}

\item{...}{Passed to methods.}
}
\value{
A data.frame of GSEA results.
}
\description{
Extract the GSEA results table
}
