% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gseaplot_with_stats}
\alias{gseaplot_with_stats}
\title{Run gseaplots and then add annotations}
\usage{
gseaplot_with_stats(gsea_obj, geneSetID, title = NULL, text_size = 6)
}
\description{
Run gseaplots and then add annotations
}
