% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{make_geneList}
\alias{make_geneList}
\title{GSEA core utilities}
\usage{
make_geneList(dt, contrast_name)
}
\description{
Functions for building ranked gene lists, running GSEA across contrasts,
retrieving specific results, and generating GSEA enrichment plots with
embedded statistics.
make a geneList from result_df from DESeq2
}
