% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{make_geneLists}
\alias{make_geneLists}
\title{make a list of geneList from combined result_df from DESeq2}
\usage{
make_geneLists(res_df)
}
\description{
make a list of geneList from combined result_df from DESeq2
}
