% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{make_gsea_matrices_dt}
\alias{make_gsea_matrices_dt}
\title{make a list mats for constructing the plotting SE object}
\usage{
make_gsea_matrices_dt(
  gsea_df,
  value_cols = c("NES", "qvalue"),
  row_col = "ID",
  col_col = "Cluster"
)
}
\description{
make a list mats for constructing the plotting SE object
}
