% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_case-class.R
\name{new_gsea_case}
\alias{new_gsea_case}
\title{Create a gsea_case object}
\usage{
new_gsea_case(deg, gsea, cell_type, pathway, contrast = "OvsY",
  cohort = "aging")
}
\arguments{
\item{deg}{Differential expression results.}

\item{gsea}{A \code{gseaResult} object.}

\item{cell_type}{Cell type label.}

\item{pathway}{Pathway (gene set) identifier.}

\item{contrast}{Contrast label.}

\item{cohort}{Cohort label.}
}
\value{
An object of class \code{gsea_case}.
}
\description{
Create a gsea_case object
}
