% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{run_pseudobulk_deseq}
\alias{run_pseudobulk_deseq}
\title{DESeq2 pseudobulk utilities}
\usage{
run_pseudobulk_deseq(
  se,
  design_formula = ~age,
  contrasts = list(O_vs_Y = c("age", "O", "Y"), O_vs_M = c("age", "O", "M"), M_vs_Y =
    c("age", "M", "Y")),
  min_count = 10,
  min_samples = 3
)
}
\description{
Functions for running pseudobulk DESeq2 on SummarizedExperiment objects
and preparing contrast-level differential expression tables for downstream
GSEA and pathway analysis.
run  DESeq2 on pseudobulk data from single cell analysi
}
